filestr: common의 service 하나로 통합
