--- helper-core/src/main/java/org/example/common/security/service/SecurityServiceImpl.java
@@ -1,14 +1,0 @@
-package org.example.common.security.service;
-
-import lombok.RequiredArgsConstructor;
-import org.example.common.security.spi.SecurityPort;
-
-@RequiredArgsConstructor
-public class SecurityServiceImpl implements SecurityService {
-    private final SecurityPort securityPort;
-
-    @Override
-    public String getCurrentUserId() {
-        return securityPort.getCurrentUserId();
-    }
-}
