+++ helper-core/src/main/java/org/example/common/service/SecurityService.java
@@ -1,4 +1,4 @@
-package org.example.common.security.service;
+package org.example.common.service;
 
 import org.springframework.stereotype.Service;
 
@@ -6,4 +6,6 @@ import org.springframework.stereotype.Service;
 public interface SecurityService {
 
     String getCurrentUserId();
+
+    void checkPasswordMatches(String rawPassword, String encodedPassword);
 }
